backend/requirements.txt
# LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0

# LangSmith
langsmith>=0.2.0

# CopilotKit + AG-UI
copilotkit>=0.1.0
ag-ui-langgraph>=0.0.20

# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Data validation
pydantic>=2.0.0

# Environment
python-dotenv>=1.0.0

# Database / Checkpointing
aiosqlite>=0.20.0
langgraph-checkpoint-sqlite>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
